[CEP] Restructures the project to match with the main repo.
