+++ testing_reports/functional/infrastructure/docker/Dockerfile
