+++ testing_reports/functional/infrastructure/docker/config/test_runner.sh
